+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     * Changes on checkout page logic - when client change the billing address just reload the SDK with the new country, but do not update the Order. The Order will be updated later.
     * Allow plugin Title settings to be set per site view.
     * Fixed some visual bugs on the checkout page.
+    * When total amount of the Order is 0, force transactionType to Auth.
 ```
 
 # 2.0.1